Nix.Utils: clean-up; add Nix.Unused
Moved function got some alteration. Reduced Lazy -> Strict marchalling.
Also instead of Lazy used Strict data types, since on the conveyor that data
seems to be consumed fully (which I may be wrong about).
